refactor(api): move /api/admin/certification-centers/{id} route
in organizational-entities
