+++ api/tests/acceptance/application/certification-centers/index_test.js
@@ -3,7 +3,6 @@ import {
   databaseBuilder,
   expect,
   generateValidRequestAuthorizationHeader,
-  insertUserWithRoleSuperAdmin,
 } from '../../../test-helper.js';
 
 describe('Acceptance | Route | Certification Centers', function () {
@@ -11,48 +10,6 @@ describe('Acceptance | Route | Certification Centers', function () {
 
   beforeEach(async function () {
     server = await createServer();
-  });
-
-  describe('PATCH /api/admin/certification-centers/{id}', function () {
-    context('when an admin member updates a certification center information', function () {
-      it('it should return an HTTP code 200 with the updated data', async function () {
-        // given
-        const adminMember = await insertUserWithRoleSuperAdmin();
-        const certificationCenter = databaseBuilder.factory.buildCertificationCenter();
-
-        await databaseBuilder.commit();
-
-        // when
-        const { result, statusCode } = await server.inject({
-          headers: {
-            authorization: generateValidRequestAuthorizationHeader(adminMember.id),
-          },
-          method: 'PATCH',
-          payload: {
-            data: {
-              id: certificationCenter.id,
-              attributes: {
-                'data-protection-officer-first-name': 'Justin',
-                'data-protection-officer-last-name': 'Ptipeu',
-                'data-protection-officer-email': 'justin.ptipeu@example.net',
-                'is-v3-pilot': true,
-                name: 'Justin Ptipeu Orga',
-                type: 'PRO',
-              },
-            },
-          },
-          url: `/api/admin/certification-centers/${certificationCenter.id}`,
-        });
-
-        // then
-        expect(statusCode).to.equal(200);
-        expect(result.data.attributes['data-protection-officer-first-name']).to.equal('Justin');
-        expect(result.data.attributes['data-protection-officer-last-name']).to.equal('Ptipeu');
-        expect(result.data.attributes['data-protection-officer-email']).to.equal('justin.ptipeu@example.net');
-        expect(result.data.attributes['is-v3-pilot']).to.equal(true);
-        expect(result.data.attributes.name).to.equal('Justin Ptipeu Orga');
-      });
-    });
   });
 
   describe('GET /api/certification-centers/{certificationCenterId}/divisions', function () {
