+++ api/tests/organizational-entities/acceptance/application/certification-center/certification-center.admin.route.test.js
@@ -6,12 +6,12 @@ import {
   insertUserWithRoleSuperAdmin,
 } from '../../../../test-helper.js';
 
-describe('Acceptance | Organization Entities | Route | Certification Centers', function () {
-  let server, request;
+describe('Acceptance | Organization Entities | Admin | Route | Certification Centers', function () {
+  let adminMember, request, server;
 
   beforeEach(async function () {
     server = await createServer();
-    await insertUserWithRoleSuperAdmin();
+    adminMember = await insertUserWithRoleSuperAdmin();
   });
 
   describe('GET /api/admin/certification-centers', function () {
@@ -356,6 +356,47 @@ describe('Acceptance | Organization Entities | Route | Certification Centers', f
 
         // then
         expect(response.statusCode).to.equal(401);
+      });
+    });
+  });
+
+  describe('PATCH /api/admin/certification-centers/{id}', function () {
+    context('when an admin member updates a certification center information', function () {
+      it('returns an HTTP code 200 with the updated data', async function () {
+        // given
+        const certificationCenter = databaseBuilder.factory.buildCertificationCenter();
+
+        await databaseBuilder.commit();
+
+        // when
+        const { result, statusCode } = await server.inject({
+          headers: {
+            authorization: generateValidRequestAuthorizationHeader(adminMember.id),
+          },
+          method: 'PATCH',
+          payload: {
+            data: {
+              id: certificationCenter.id,
+              attributes: {
+                'data-protection-officer-first-name': 'Justin',
+                'data-protection-officer-last-name': 'Ptipeu',
+                'data-protection-officer-email': 'justin.ptipeu@example.net',
+                'is-v3-pilot': true,
+                name: 'Justin Ptipeu Orga',
+                type: 'PRO',
+              },
+            },
+          },
+          url: `/api/admin/certification-centers/${certificationCenter.id}`,
+        });
+
+        // then
+        expect(statusCode).to.equal(200);
+        expect(result.data.attributes['data-protection-officer-first-name']).to.equal('Justin');
+        expect(result.data.attributes['data-protection-officer-last-name']).to.equal('Ptipeu');
+        expect(result.data.attributes['data-protection-officer-email']).to.equal('justin.ptipeu@example.net');
+        expect(result.data.attributes['is-v3-pilot']).to.equal(true);
+        expect(result.data.attributes.name).to.equal('Justin Ptipeu Orga');
       });
     });
   });
